Move WindowIdExt to its own file; eliminate force-paint hack - better fix available
