+++ src/lib.rs
@@ -181,6 +181,7 @@ pub mod view_tuple;
 pub mod views;
 pub mod window;
 mod window_handle;
+mod window_id;
 mod window_tracking;
 
 pub use app::{launch, quit_app, AppEvent, Application};
@@ -189,10 +190,11 @@ pub use clipboard::{Clipboard, ClipboardError};
 pub use floem_reactive as reactive;
 pub use floem_renderer::cosmic_text;
 pub use floem_renderer::Renderer;
-pub use id::{ViewId, WindowIdExt};
+pub use id::ViewId;
 pub use peniko;
 pub use peniko::kurbo;
 pub use screen_layout::ScreenLayout;
 pub use taffy;
 pub use view::{recursively_layout_view, AnyView, IntoView, View};
 pub use window::{close_window, new_window};
+pub use window_id::WindowIdExt;
